Sunburst Chart fixed
My goal was to be able, from any project that uses the
plugin, to create the appropriate data and set them
on the Sunburst chart.

After I created an App\Services\SunburstInstantiationService
that does the aforementioned task, I set the plugin
to receive these data and then pass them to the
chart view. Among other things, that was done by declaring on the
SunburstChart class, that extends Widget, a public
data variable that is accessible from the view.
That variable is then used by the Javascript method
SunburstObject to create the Sunburst Chart.

Other tasks at hand are the manual configuration
of the chart description, the use of bright colors,
the tinkering of the chart style and size and the
ability of the service, that creates the data, to
receive a Closure|array of data (regardless of type
restriction) and produce a chart.