Failed chart declaration autonomy
I did some changes on the project's sunburst instantiation service so that the data of the chart and the information describing on the title and description are passed and used correctly from the blade view.

However, the plugin will be unable to work after this commit, since I attempted to make a widget that can be declared from anywhere in the main project with the apropriate values but with no success(thus the changes on adminpanelprovider, SunburstChart and MultipieChart).

I am putting this branch on hold as I will start from scratch on creating a plugin that when installed and called with a command, it will create the appropriate classes and methods required for the instantiation of a sunburst chart.