Issue solved & Plugin Stub and Command addition
(1): I wanted to find a way to pass data from
my widget class to the blade view. By declaring
public variables of these data and instantiating them
(which I do, in turn, in the mount method),
the blade view can get access to them.

(2): I wanted anyone who would use the plugin to install
it and run a command that would manage the nessesary
setup. Then, the user would provide the plugin with
his data and that would do the rest for rendering it.
To do so:
    (2.1): I turned the SunburstChart widget class into
    an abstract one by declaring a method getData().
    Therefore, any class that extends the abstract widget
    class will have to implement it. The getData in turn will
    instantiate the respective variables so that the
    blade view can have access to them.
    (2.2): I created a command that gets a .stub
    file from the plugin package and creates a widget
    class in the App\\Filament\\Widgets folder according
    to it.
    (2.3): Considering the (2.2), I created a stub
    file in tsetis/sunburst-chart/stubs folder that
    has a template class for context. This class
    extends the SunburstChart widget class in order
    for the user to use the getData() method and
    therefore instantiate the chart.